fix: absolute position issues
Fixes two issues with absolute position:

1. Text disappears when text is split into lines only, and target
element does not contain any nested elements.

2. Splitting text with absolute position slightly changes the visual
appearance of the text. This was due to the fact that we used
`offsetTop` and `offsetLeft`, which do not have sub-pixel precision.
Now, we use `getBoundingClientRect` to determine the size and
position of split text nodes. This has a slight negative impact on
performance. But it prevents text from visibly shifting when it
is split/un-split.