fix: absolute position issues (#30)
Fixes two issues related to absolute position:

1. Text disappears when text is only split into lines, and the target
element does not contain any nested elements. This was a bug 
introduced by #28. 

2. Splitting text with absolute position slightly changes the visual
appearance of the text. This was due to the fact that we used
`offsetTop` and `offsetLeft`, which do not have sub-pixel precision.
Now, we use `getBoundingClientRect` to determine the position
of split text nodes. This has a slight negative impact on performance
(when absolute position is enabled), but it prevents the text from
from visibly shifting around when it is split/un-split.