fix issue 47
Issue: if you call `instance.split` multiple times, the html content of
the target element(s) is replaced with "undefined"

This was due to fact that all stored data associated with the elements
is removed in the `revert` method. So after reverting the elements, the
could not be split again (this bug was introduced in #41).

Fix: store the original html content of the elements at the beginning
of the split method instead of the constructor.